#63 Add possibility to skip ssh
When running against an RDS, we wont have ssh connection to the instance. Added option to skip ssh checks, this way $can_run_os_cmd will be always false.

Since we don't have ssh, the script won't be able to search for total_memory. Added a new parameter for the total memory of the database to be manually informed. This way we can still have the memory ratio warnings.